Move audit into settings sections
